+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- ChilloutVR world and group details no longer blink away when their five-minute background cache refreshes. (VRX-265)
+
 ## [0.18.1] - 2026-08-14
 
 ### Fixed